Types dedicated to 3DS sdk

+++ src/vision.cpp
@@ -136,12 +136,12 @@ void draw_bboxes(const cv::Mat& image, const std::vector<BoxInfo>& bboxes, int v
         {
             struct 
             {
-                uint8_t r;
-                uint8_t g;
-                uint8_t b;
-                uint8_t a;
+                u8 r;
+                u8 g;
+                u8 b;
+                u8 a;
             };
-            int rgba;
+            u32 rgba;
         } color;
 
         color.r = color_list[bbox.label][0];
@@ -160,7 +160,7 @@ void draw_bboxes(const cv::Mat& image, const std::vector<BoxInfo>& bboxes, int v
             (int) y1,
             (int) w,
             (int) h,
-            color.rgba, 
+            (int) color.rgba, 
             2
         );
 
@@ -172,7 +172,7 @@ void draw_bboxes(const cv::Mat& image, const std::vector<BoxInfo>& bboxes, int v
             (int) x1 + 1, 
             (int) y1 + 1,
             7,
-            color.rgba
+            (int) color.rgba
         );
     }
     printf("=======================================\n");
